% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_parameters.R
\name{GetPostEta}
\alias{GetPostEta}
\title{Draw parameters from substantive model}
\usage{
GetPostEta(t, x, v, m = 5, fixed = TRUE)
}
\arguments{
\item{t}{a tibble with one column (t) of effect estimates}

\item{x}{a tibble of covariates}

\item{v}{a tibble with one column (v) of estimation variances}

\item{m}{the number of posterior draws to generate}

\item{fixed}{boolean indicates fixed/random effects model}
}
\value{
tibble of posterior draws from a MR model
}
\description{
Draw parameters from substantive model
}
