% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_parameters.R
\name{GetPostPsi}
\alias{GetPostPsi}
\title{Draw parameters from partial imputation model}
\usage{
GetPostPsi(x, v, variable, m = 5, type = NULL)
}
\arguments{
\item{x}{a tibble of covariates}

\item{v}{a tibble with one column (v) of variances}

\item{variable}{a string indicating which variable in \code(x) to model}

\item{m}{a number of draws from the posterior to make}

\item{type}{string indicating 'binary' or 'continuous'}
}
\value{
tibble of draws from auxiliary model
}
\description{
Draw parameters from partial imputation model
}
