% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_x.R
\name{ImputeX}
\alias{ImputeX}
\title{Draw new values of a single covariate (given other covariates)}
\usage{
ImputeX(t, x, v, eta, psi, variable, type)
}
\arguments{
\item{t}{a tibble of effect sizes (one column called 't')}

\item{x}{a tibble of covariates}

\item{v}{a tibble of estimation variances (one column called 'v')}

\item{eta}{tibble of draws from the posterior of the MR model}

\item{psi}{tibble of draws from the aux model}

\item{variable}{string indicating which covariate to impute}

\item{type}{string indicating covariate type ('binary', 'continuous')}
}
\value{
tibble with imputed values for \code(variable) in \code(x)
}
\description{
Draw new values of a single covariate (given other covariates)
}
