% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{Initialize}
\alias{Initialize}
\title{Initialize dataset for chained imputations}
\usage{
Initialize(x)
}
\arguments{
\item{x}{a tibble of covariates}
}
\value{
the same tibble \code(x) with missing values filled in
}
\description{
Initialize dataset for chained imputations
}
