% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi.R
\name{MetaMice}
\alias{MetaMice}
\title{Generate imputations for a multiple covariates}
\usage{
MetaMice(t, x, v, m = 5, variables, types, fixed = TRUE, burn_in = 10)
}
\arguments{
\item{t}{a tibble of effect sizes (one column called 't')}

\item{x}{a tibble of covariates}

\item{v}{a tibble of estimation variances (one column called 'v')}

\item{m}{number of imputations to generate}

\item{variables}{list of strings indicating which variables in \code(x) to impute}

\item{types}{list of strings indicating covariate types ('binary', 'continuous')}

\item{fixed}{boolean indicates fixed/random effects model}

\item{burn_in}{number of burn-in iterations for chained equations}
}
\value{
mids object of imputed datasets
}
\description{
Generate imputations for a multiple covariates
}
